Refactor magento script, add sources, minor fixes

+++ shellscripts/magento.sh
@@ -10,12 +10,71 @@ if [[ $(/usr/bin/id -u) -ne 0 ]]; then
     echo "Not running as root"
     exit
 fi
+  
+# Check for argument presence
+    if [[ $1 = "--help" || $1 = "" ]]; then
+        echo "Usage:"
+        echo "  --help          run this help"
+        echo " -c --create-db     database creation"
+        echo " -d --download      download magento"
+        echo " -p --permissions    Set permissions"
+        echo " -a --apache-setup  basic apache setup"
+        echo " -i --install       install magento"
+    fi
 
+    argument_lists=( "$@" )
+    for i in "${argument_lists[@]}"; do
+      if [[ $i = "--create-db" ]] ; then
+        CREATE_DB="yes"
+      elif [[ $i = "--apache-setup" ]]; then
+        APACHE_SETUP="yes" 
+      elif [[ $i = "--create-db" ]] ; then
+        DOWNLOAD="yes"
+      elif [[ $i = "--permissions" || $i = "-p" ]] ; then
+        PERMISSIONS="yes"
+      elif [[ $i = "--install" ]]; then
+        INSTALL="yes"
+      elif [[ $i = "--ttt" ]]; then
+        TTT="yes"
+      fi 
+    done
+
+if [[ $APACHE_SETUP = "yes" ]]; then
+  # Delete default apache web dir and symlink mounted vagrant dir from host machine
+  rm -rf /var/www
+  mkdir /vagrant/httpdocs
+  ln -fs /vagrant/httpdocs /var/www
+
+  # Replace contents of default Apache vhost
+  # --------------------
+VHOST=$(cat <<EOF
+<VirtualHost *:80>
+  DocumentRoot "/var/www"
+  ServerName localhost
+  <Directory "/var/www">
+    Options FollowSymLinks  
+    AllowOverride All
+  </Directory>
+</VirtualHost>
+EOF
+)
+
+  echo "$VHOST" > /etc/apache2/sites-enabled/000-default
+
+# Restarting services
+  a2enmod rewrite
+  service apache2 restart
+fi
+
+
+if [[ $CREATE_DB = "yes" ]]; then
 # Create database
-  mysql -u root -p root -e "CREATE DATABASE IF NOT EXISTS boxdatabase"
-  mysql -u root -p root -e "GRANT ALL PRIVILEGES ON boxdatabase.* TO 'boxuser'@'localhost' IDENTIFIED BY 'boxpassword'"
-  mysql -u root -p root -e "FLUSH PRIVILEGES"
+  mysql -uroot -proot -e"CREATE DATABASE IF NOT EXISTS boxdatabase"
+  mysql -uroot -proot -e"GRANT ALL PRIVILEGES ON boxdatabase.* TO 'boxuser'@'localhost' IDENTIFIED BY 'boxpassword'"
+  mysql -uroot -proot -e"FLUSH PRIVILEGES"
+fi
 
+if [[ $DOWNLOAD = "yes" ]]; then
 # Download and extract
 if [[ ! -f "/vagrant/httpdocs/index.php" ]]; then
   if [[ ! -d "/vagrant/httpdocs" ]]; then
@@ -25,12 +84,24 @@ if [[ ! -f "/vagrant/httpdocs/index.php" ]]; then
   wget http://www.magentocommerce.com/downloads/assets/1.7.0.2/magento-1.7.0.2.tar.gz
   tar -zxvf magento-1.7.0.2.tar.gz
   mv magento/* magento/.htaccess .
+fi
+fi
+
+if [[ $PERMISSIONS = "yes" ]]; then
+  find . -type f -exec chmod 644 {} \;
+  find . -type d -exec chmod 755 {} \;
   chmod -R o+w media var
   chmod o+w app/etc
+  chmod 550 pear
+fi
+
+if [[ $DOWNLOAD = "yes" ]]; then
   # Clean up downloaded file and extracted dir
   rm -rf magento*
 fi
 
+
+if [[ $INSTALL = "yes" ]]; then
 # Run installer
 if [[ ! -f "/vagrant/httpdocs/app/etc/local.xml" ]]; then
   cd /vagrant/httpdocs
@@ -43,3 +114,9 @@ if [[ ! -f "/vagrant/httpdocs/app/etc/local.xml" ]]; then
   --admin_lastname Owner --admin_firstname Store --admin_email "admin@example.com" \
   --admin_username admin --admin_password password123
 fi
+fi
+
+if [[ $TTT ]]; then
+    echo Works
+fi
+
